Added module to isolate functions using the template engine.
