+++ erddapcfg/utils.py
@@ -1,8 +1,5 @@
 from enum import StrEnum
 
-from jinja2 import Environment, PackageLoader, select_autoescape
-from .classes import ERDDAP
-
 
 class Mode(StrEnum):
     """Enum used to select the mode of the line ending conversion."""
@@ -36,80 +33,3 @@ def change_line_ending(filename: str, mode: Mode = Mode.Win2Unix):
     # Save converted file
     with open(filename, "wb") as open_file:
         open_file.write(content)
-
-
-def obj2sqlscript(erddap: ERDDAP):
-    output = []
-
-    # Render the template
-    env = Environment(loader=PackageLoader("erddapcfg"), autoescape=select_autoescape())
-    template = env.get_template("db_insert.j2")
-    output.append(
-        template.render(
-            params=erddap.params,
-            datasets=erddap.datasets,
-            dataset_params=[
-                {"name": param.name, "value": param.value, "datasetID": dataset.datasetID}
-                for dataset in erddap.datasets
-                for param in dataset.params
-            ],
-            dataset_attributes=[
-                {
-                    "name": attribute.name,
-                    "type": attribute.type,
-                    "text": attribute.text,
-                    "datasetID": dataset.datasetID,
-                }
-                for dataset in erddap.datasets
-                for attribute in dataset.attributes
-            ],
-            dataset_source_attributes=[
-                {
-                    "name": attribute.name,
-                    "type": attribute.type,
-                    "text": attribute.text,
-                    "datasetID": dataset.datasetID,
-                }
-                for dataset in erddap.datasets
-                for attribute in dataset.source_attributes
-            ],
-            variables=[
-                {
-                    "tag": variable.tag,
-                    "sourceName": variable.sourceName,
-                    "destinationName": variable.destinationName,
-                    "dataType": variable.dataType,
-                    "datasetID": dataset.datasetID,
-                }
-                for dataset in erddap.datasets
-                for variable in dataset.variables
-            ],
-            variable_attributes=[
-                {
-                    "name": attribute.name,
-                    "type": attribute.type,
-                    "text": attribute.text,
-                    "destinationName": variable.destinationName,
-                    "datasetID": dataset.datasetID,
-                }
-                for dataset in erddap.datasets
-                for variable in dataset.variables
-                for attribute in variable.attributes
-            ],
-            variable_source_attributes=[
-                {
-                    "name": attribute.name,
-                    "type": attribute.type,
-                    "text": attribute.text,
-                    "destinationName": variable.destinationName,
-                    "datasetID": dataset.datasetID,
-                }
-                for dataset in erddap.datasets
-                for variable in dataset.variables
-                for attribute in variable.source_attributes
-            ],
-            dataset_children=erddap.parent_child,
-        )
-    )
-
-    return "".join(output)
